pyproject.toml
[tool.poetry]
name = "lightning-geometric"
version = "0.1.0"
description = "Pytorch Geometric Examples using Pytorch Lightning"
authors = ["Thomas Chaton <thomas.chaton.ai@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
torch = "^1.6.0"
hydra-core = "1.0.0"
torch-cluster = "^1.5.7"
torch-sparse = "^0.6.7"
torch-scatter = "^2.0.5"
torch-geometric = "^1.6.1"
pytorch-lightning = { git = "https://github.com/PyTorchLightning/pytorch-lightning.git", branch = "master" }
openmesh = "^1.1.4"
torch-spline-conv = "^1.2.0"
ogb = "^1.2.3"
wandb = "^0.10.2"

[tool.poetry.dev-dependencies]
black = {version = "^20.8b1", allow-prereleases = true}

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
